requirements.txt
Django==1.9.6
django-bootstrap3==7.0.1
django-bootstrap-ui==0.2.0
django-notifications-hq